style: 요구사항에 맞게 소스파일 분리 및 코드 리팩토링
